Add an option for not following redirects
Adds a new parameter `follow_redirects` which can be set to `False`
to turn off automatic redirect following, so that the original 3xx response
is returned. This allows the user to implement their own redirect logic - for
example to inspect the Location header before following a redirect.

Default value is `True` which preserves the existing
behavior.