close socket after sending request with `connection: close` header
Currently when 'asks' uses a session without `connection: keep-alive` it
relies on the server to send the header `connection: close` back. So
even if 'asks' itself asks the server to close the connection it will
try to reuse the connection in the session unless the server sends that
header back. This is incorrect since it says in the official W3 standard

https://www.w3.org/Protocols/rfc2616/rfc2616-sec8.html
Once a close has been signaled, the client MUST NOT send any more
requests on that connection.

So 'asks' should close the socket after receiving a response.
Closes #115